Don't include more than one blank option in conditional logic dropdown

+++ classes/models/FrmFieldOption.php
@@ -69,11 +69,13 @@ class FrmFieldOption {
 	 * @param int $truncate
 	 */
 	public function print_single_option( $selected_value, $truncate ) {
-		echo '<option value="' . esc_attr( $this->saved_value ) . '"';
-		selected( esc_attr( $selected_value ), esc_attr( $this->saved_value ) );
-		// TODO: add hook that can add attributes to option text
-		echo '>';
-		echo FrmAppHelper::truncate( $this->option_label, $truncate ) . '</option>';
+		if ( $this->saved_value !== '' ) {
+			echo '<option value="' . esc_attr( $this->saved_value ) . '"';
+			selected( esc_attr( $selected_value ), esc_attr( $this->saved_value ) );
+			// TODO: add hook that can add attributes to option text
+			echo '>';
+			echo FrmAppHelper::truncate( $this->option_label, $truncate ) . '</option>';
+		}
 	}
 
 }
